change Fortran keywords to lower-case letters
Following a recommendation,[1] keywords of Fortran in the snippets
of code, running text, and section headers now are set in lower-
case characters only.  Beside syntax highlighlighting the web site
offers, this then is more consistent in style with the other booklets
already present.  There however are exceptions to this change in the
table about value statements, section Performing I/O with objects of
container-like type).

[1] https://github.com/fortran-lang/webpage/pull/477

Signed-off-by: Norwid Behrnd <nbehrnd@yahoo.com>